feat: CNS-891: Same provider conflict (#1614)
* Formatting and typos

* Change the FinalizationConflict relayReply message type and use it instead

* Add block_distance_to_finalization and spec_id to RelayFinalization

* Move utils function to correct file

* Implement the ValidateSameProviderConflict func

* Add more validations

* Small rename

* Fix bugs, made renames and made some logs clearer

* Add conflict tests

* Remove the relay session from the RelayFinalization

* Some code cleaning

* Move VerifyFinalizationData to finalization_consensus

* Cancel Conflict Detection Tx on finalizationAccountabilityError for single reply

* Send proof with finalization error

* Change conflict proto to use oneof

* Remove TODO

* Move FindLargestIntValueInMap to maps.go

* Allow different epoch in finalization in consensus

* Add Stringify to RelayFinalization

* Fix a small bug in maps

* Made fixes to the finalization consensus

* Fix lint

* Small fix to construct types

* Changed the proto once again, and fixed some tests

* Post-merge fixes

* Remove SameProviderConflict from proto for better readability

* Small fixes

* Remove old code

* Small fixes

* Important fixes to the finalization

* Add some tests to the same provider conflict

* Comment out the corresponding SDK code

* Fix log

* Fix lint

* Update README

* CR Fix: Implement TODO - verify finalization block hashes length

* CR Fix: remove irrelevant TODO

* CR Fix: lier -> liar

* CR Fix: Rename test

* CR Fix: Revert TODO comment delete

* CR Fix: Remove resolved TODO

* CR Fix: Add test case

* CR Fix: Rename

* Update test timeout

* Update test timeout again

* CR Fix: Tiny rename

* CR Fix: RelayReply -> RelayFinalization

* CR Fix: Small renames

* CR Fix: Remove TODO

Approved by the @omerlavanet

* Remove unused returned struct

* Fix event on conflict not showing provider address

* Add some checks

* CR Fix: Use less expensive function to check for finalization of block

* CR Fix: Remove usage of float64 in consensus

* Remove old, unused code

Requested by @yaroms

* make it better

* Fix broken fix

* CR Fix: Restore the atomic store

* CR Fix: Rename proto names to be snake case

* CR Fix: Rename currentBlockToHashesToAgreeingProviders -> currentEpochBlockHashesToAgreeingProviders

* CR Fix: Make the code a little more clear around locks

* Remove comment

* CR Fix: Rename lock name

* Use make for creating new maps

* CR Fix: Add comment to describe function

* Update protocol/rpcprovider/rpcprovider_server.go


* CR Fix: Remove math.floor and add tests to Median func

* Remove all "if debug", and replace them with Trace logs instead

* Add some trace logs to finalization conflict + minor changes

* Set the globalLogLevel back to debug and rename it

* Fix lint

* Add comment that says what are the keys

* CR Fix: Remove done TODO

* CR Fix: Median loss of precision + test

* CR Fix: Add comment to explain why we use sort function

* CR Fix: Add some more comments

* Post merge fixes

* Squashed 'CNS-891-same-provider-conflict-benchmark-test' changes

* Lint fix

* Remove old errors.go file

* Fix finalization test

* fixing test races

---------

+++ protocol/chainlib/chain_fetcher.go
@@ -368,7 +368,7 @@ func (cf *ChainFetcher) FetchBlockHashByNum(ctx context.Context, blockNum int64)
 	_, _, blockDistanceToFinalization, _ := cf.chainParser.ChainBlockStats()
 	latestBlock := atomic.LoadInt64(&cf.latestBlock) // assuming FetchLatestBlockNum is called before this one it's always true
 	if latestBlock > 0 {
-		finalized := spectypes.IsFinalizedBlock(blockNum, latestBlock, blockDistanceToFinalization)
+		finalized := spectypes.IsFinalizedBlock(blockNum, latestBlock, int64(blockDistanceToFinalization))
 		isNodeError, _ := chainMessage.CheckResponseError(reply.RelayReply.Data, reply.StatusCode)
 		if !isNodeError { // skip cache populate on node errors, this is a protection but should never get here with node error as we parse the result prior.
 			cf.populateCache(cf.constructRelayData(collectionData.Type, path, data, blockNum, "", nil, latestBlock), reply.RelayReply, []byte(res), finalized)
